restructured folders and removed relative path import
