+++ react-components/src/components/RevealContext/RevealContext.tsx
@@ -7,14 +7,14 @@ import { type ReactNode, useEffect, useMemo, useRef, useState, type ReactElement
 import { type Color } from 'three';
 import { I18nContextProvider } from '../i18n/I18n';
 import { ViewerContext } from '../RevealCanvas/ViewerContext';
-import { NodeCacheProvider } from '../NodeCacheProvider/NodeCacheProvider';
-import { AssetMappingCacheProvider } from '../NodeCacheProvider/AssetMappingCacheProvider';
-import { PointCloudAnnotationCacheProvider } from '../NodeCacheProvider/PointCloudAnnotationCacheProvider';
+import { NodeCacheProvider } from '../CacheProvider/NodeCacheProvider';
+import { AssetMappingCacheProvider } from '../CacheProvider/AssetMappingCacheProvider';
+import { PointCloudAnnotationCacheProvider } from '../CacheProvider/PointCloudAnnotationCacheProvider';
 import { Reveal3DResourcesCountContextProvider } from '../Reveal3DResources/Reveal3DResourcesCountContext';
 import { SDKProvider } from '../RevealCanvas/SDKProvider';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { useRevealKeepAlive } from '../RevealKeepAlive/RevealKeepAliveContext';
-import { Image360AnnotationCacheProvider } from '../NodeCacheProvider/Image360AnnotationCacheProvider';
+import { Image360AnnotationCacheProvider } from '../CacheProvider/Image360AnnotationCacheProvider';
 
 export type RevealContextProps = {
   color?: Color;
