+++ react-components/src/components/RevealKeepAlive/RevealKeepAlive.tsx
@@ -5,10 +5,10 @@
 import { type Cognite3DViewer } from '@cognite/reveal';
 import { type ReactNode, type ReactElement, useRef, useEffect } from 'react';
 import { RevealKeepAliveContext } from './RevealKeepAliveContext';
-import { type FdmNodeCache } from '../NodeCacheProvider/FdmNodeCache';
-import { type AssetMappingCache } from '../NodeCacheProvider/AssetMappingCache';
-import { type PointCloudAnnotationCache } from '../NodeCacheProvider/PointCloudAnnotationCache';
-import { type Image360AnnotationCache } from '../NodeCacheProvider/Image360AnnotationCache';
+import { type FdmNodeCache } from '../CacheProvider/FdmNodeCache';
+import { type AssetMappingCache } from '../CacheProvider/AssetMappingCache';
+import { type PointCloudAnnotationCache } from '../CacheProvider/PointCloudAnnotationCache';
+import { type Image360AnnotationCache } from '../CacheProvider/Image360AnnotationCache';
 
 export function RevealKeepAlive({ children }: { children?: ReactNode }): ReactElement {
   const viewerRef = useRef<Cognite3DViewer>();
