+++ react-components/src/components/RevealKeepAlive/RevealKeepAliveContext.ts
@@ -3,10 +3,10 @@
  */
 import { type Cognite3DViewer } from '@cognite/reveal';
 import { type MutableRefObject, createContext, useContext } from 'react';
-import { type FdmNodeCache } from '../NodeCacheProvider/FdmNodeCache';
-import { type AssetMappingCache } from '../NodeCacheProvider/AssetMappingCache';
-import { type PointCloudAnnotationCache } from '../NodeCacheProvider/PointCloudAnnotationCache';
-import { type Image360AnnotationCache } from '../NodeCacheProvider/Image360AnnotationCache';
+import { type FdmNodeCache } from '../CacheProvider/FdmNodeCache';
+import { type AssetMappingCache } from '../CacheProvider/AssetMappingCache';
+import { type PointCloudAnnotationCache } from '../CacheProvider/PointCloudAnnotationCache';
+import { type Image360AnnotationCache } from '../CacheProvider/Image360AnnotationCache';
 
 export type RevealKeepAliveData = {
   viewerRef: MutableRefObject<Cognite3DViewer | undefined>;
