+++ react-components/src/hooks/use3dModels.ts
@@ -3,9 +3,9 @@
  */
 
 import { useMemo } from 'react';
-import { useReveal } from '..';
 import { useReveal3DResourcesCount } from '../components/Reveal3DResources/Reveal3DResourcesCountContext';
 import { type CogniteModel } from '@cognite/reveal';
+import { useReveal } from '../components/RevealCanvas/ViewerContext';
 
 export const use3dModels = (): CogniteModel[] => {
   const viewer = useReveal();
