+++ react-components/src/hooks/useSceneDefaultCamera.tsx
@@ -3,11 +3,11 @@
  */
 
 import { useMemo } from 'react';
-import { useSceneConfig } from './useSceneConfig';
+import { useSceneConfig } from '../query/useSceneConfig';
 import { Vector3, Quaternion, Euler, MathUtils, Box3 } from 'three';
-import { useReveal } from '..';
 import { CDF_TO_VIEWER_TRANSFORMATION, type Cognite3DViewer } from '@cognite/reveal';
 import { type SceneConfiguration } from '../components/SceneContainer/SceneTypes';
+import { useReveal } from '../components/RevealCanvas/ViewerContext';
 
 export const useSceneDefaultCamera = (
   sceneExternalId: string | undefined,
