+++ react-components/src/query/useSceneConfig.ts
@@ -20,7 +20,7 @@ import {
 } from '../components/SceneContainer/SceneTypes';
 import { useFdmSdk } from '../components/RevealCanvas/SDKProvider';
 import { type Source, type FdmSDK } from '../utilities/FdmSDK';
-import { type SceneConfigurationProperties } from './types';
+import { type SceneConfigurationProperties } from '../hooks/types';
 import { fdmViewsExist } from '../utilities/fdmViewsExist';
 
 const DefaultScene: Scene = {
