+++ react-components/src/query/useSearchAssetsMappedPointCloudAnnotations.ts
