+++ react-components/src/query/useSearchMappedEquipmentFDM.tsx
